Move SlotCommitmentId to slot.py
