+++ bindings/python/iota_sdk/__init__.py
@@ -43,6 +43,7 @@ from .types.output_metadata import *
 from .types.output_params import *
 from .types.payload import *
 from .types.send_params import *
+from .types.slot import *
 from .types.token_scheme import *
 from .types.transaction_with_metadata import *
 from .types.transaction_data import *
