+++ bindings/python/iota_sdk/types/output_metadata.py
@@ -7,9 +7,9 @@ from dataclasses import dataclass, field
 from dataclasses_json import config
 from iota_sdk.types.block.block import BlockId
 from iota_sdk.types.common import SlotIndex, json
-from iota_sdk.types.context_input import SlotCommitmentId
 from iota_sdk.types.output import AccountOutput, BasicOutput, DelegationOutput, FoundryOutput, NftOutput, deserialize_output
 from iota_sdk.types.output_id import OutputId
+from iota_sdk.types.slot import SlotCommitmentId
 from iota_sdk.types.transaction_data import TransactionId
