+++ bindings/python/iota_sdk/types/slot.py
@@ -1,0 +1,9 @@
+# Copyright 2024 IOTA Stiftung
+# SPDX-License-Identifier: Apache-2.0
+
+from iota_sdk.types.common import IdWithSlotIndex
+
+
+class SlotCommitmentId(IdWithSlotIndex):
+    """Represents a slot commitment ID.
+    """
